added token authentication and finished the exercises

+++ controllers/blogs.js
@@ -1,9 +1,11 @@
 const blogsRouter=require('express').Router()
 const Blog=require('../models/blog')
+const User=require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response,next) => {
    try{
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user',{username:1,name:1})
     console.log(blogs)
     if(blogs){
         response.json(blogs.map(blog=>blog.toJSON()))
@@ -16,14 +18,20 @@ blogsRouter.get('/', async (request, response,next) => {
   })
   
   blogsRouter.post('/', async (request, response,next) => {
-    const blog = new Blog(request.body)
+    const body = request.body
+    const token= request.token
     try{
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }   
+        const user = await User.findById(decodedToken.id)
+        body.user=user._id
+        const blog=new Blog(body)
         const result= await blog.save()
-        if(result){
-            response.status(201).json(result.toJSON())
-        }else{
-            response.status(404).end()
-        }
+        user.blogs=user.blogs.concat(result._id)
+        await user.save()
+        response.json(result.toJSON())
     }catch(err){
         next(err)
     }
@@ -31,9 +39,20 @@ blogsRouter.get('/', async (request, response,next) => {
 
   blogsRouter.delete('/:id', async (request,response,next) => {
     const id = request.params.id
+    const token= request.token
     try{
-        await Blog.findByIdAndRemove(id);
-        response.status(204).end()
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const blog=await Blog.findById(id)
+        if(blog.user.toString()===decodedToken.id) {
+            await Blog.findByIdAndRemove(id);
+            return response.status(204).end()
+        }else{
+            return response.status(400).json({error : 'not authorized to delete the post'})
+        }  
+       
     }catch(err){
         next(err)
     }
